requirements.txt
streamlit
pandas
seaborn
matplotlib
google-generativeai
python-dotenv
edgartools
json